Edited Methods to work on DAG and misc
Changed findPath function to work with a successor List as opposed to the left and right pointers that we used for the binary tree Implementation. Created New findLCADAG function that backtracks from n1 and n2 and attempts to find Lowest Common Ancestor of the Nodes. This implementation currently does not work with many cases.
Misc: Added Node constructor.